+++ src/components/Form/InputField/InputField.stories.tsx
@@ -10,7 +10,8 @@ Text.args = {
     type: "text",
     label: "First Name",
     placeholder: "Enter your name",
-    helpText: "This field can only contain a string"
+    helpText: "This field can only contain a string",
+    errorText: "Looks invalid, re-check?"
 };
 
 export const Number: FictoanStory<typeof InputField> = createStoryFromTemplate(Template);
@@ -18,7 +19,8 @@ Number.args = {
     type: "number",
     label: "Age",
     placeholder: "Enter your age",
-    helpText: "This field can only contain a number"
+    helpText: "This field can only contain a number",
+    errorText: "Looks invalid, re-check?"
 };
 
 export const Email: FictoanStory<typeof InputField> = createStoryFromTemplate(Template);
@@ -26,12 +28,14 @@ Email.args = {
     type: "email",
     label: "Email",
     placeholder: "Enter your email address",
-    helpText: "This field can only contain an email"
+    helpText: "This field can only contain an email",
+    errorText: "Looks invalid, re-check?"
 };
 
 export const Password: FictoanStory<typeof InputField> = createStoryFromTemplate(Template);
 Password.args = {
     type: "password",
     label: "Password",
-    helpText: "This field will obfuscate your input"
+    helpText: "This field will obfuscate your input",
+    errorText: "Looks invalid, re-check?"
 };